List `scipy` in `requirements.txt`
The `scipy` package was not listed in the `requirements.txt`. This is a non-issue in terms of the Docker Image
as `scipy` is installed with the `scikit-survival` package.
However, when setting up a new workspace, listing it in the requirements is beneficial for the code insight features of the IDE.