Add allow censored value to be given
We take a value for the event if it occured and when censored. The event value is handled as a string type now.
The event array is checked for fields where the event occured, and where the event is censored.
When a value is given that is neither the row is dropped in both (censoring state and time) arrays and the feature DataFrame.